requirements.txt
asgiref==3.11.0
click==8.3.1
colorama==0.4.6
dj-database-url==3.1.0
Django==5.0.4
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==24.1.1
h11==0.16.0
packaging==26.0
psycopg2-binary==2.9.11
sqlparse==0.5.5
tzdata==2025.3
uvicorn==0.40.0
whitenoise==6.11.0